Updated ThresholdDetectionProcessor to new API
