--- streampipes-processors-filters-jvm/src/main/java/org/apache/streampipes/processors/filters/jvm/processor/threshold/ThresholdDetectionController.java
@@ -1,92 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.apache.streampipes.processors.filters.jvm.processor.threshold;
-
-import org.apache.streampipes.model.DataProcessorType;
-import org.apache.streampipes.model.graph.DataProcessorDescription;
-import org.apache.streampipes.model.graph.DataProcessorInvocation;
-import org.apache.streampipes.model.schema.PropertyScope;
-import org.apache.streampipes.sdk.builder.ProcessingElementBuilder;
-import org.apache.streampipes.sdk.builder.StreamRequirementsBuilder;
-import org.apache.streampipes.sdk.extractor.ProcessingElementParameterExtractor;
-import org.apache.streampipes.sdk.helpers.*;
-import org.apache.streampipes.sdk.utils.Assets;
-import org.apache.streampipes.vocabulary.SO;
-import org.apache.streampipes.wrapper.standalone.ConfiguredEventProcessor;
-import org.apache.streampipes.wrapper.standalone.declarer.StandaloneEventProcessingDeclarer;
-
-public class ThresholdDetectionController extends StandaloneEventProcessingDeclarer<ThresholdDetectionParameters> {
-
-  private static final String NUMBER_MAPPING = "number-mapping";
-  private static final String VALUE = "value";
-  private static final String OPERATION = "operation";
-
-  private static final String RESULT_FIELD = "thresholdDetected";
-
-  @Override
-  public DataProcessorDescription declareModel() {
-    return ProcessingElementBuilder.create("org.apache.streampipes.processors.filters.jvm.threshold")
-            .category(DataProcessorType.FILTER)
-            .withAssets(Assets.DOCUMENTATION, Assets.ICON)
-            .withLocales(Locales.EN)
-            .requiredStream(StreamRequirementsBuilder
-                    .create()
-                    .requiredPropertyWithUnaryMapping(EpRequirements.numberReq(),
-                            Labels.withId(NUMBER_MAPPING),
-                            PropertyScope.NONE).build())
-            .outputStrategy(
-                    OutputStrategies.append(
-                            EpProperties.booleanEp(Labels.empty(), RESULT_FIELD, SO.Boolean)))
-            .requiredSingleValueSelection(Labels.withId(OPERATION), Options.from("<", "<=", ">",
-                    ">=", "==", "!="))
-            .requiredFloatParameter(Labels.withId(VALUE), NUMBER_MAPPING)
-            .build();
-
-  }
-
-  @Override
-  public ConfiguredEventProcessor<ThresholdDetectionParameters> onInvocation
-          (DataProcessorInvocation sepa, ProcessingElementParameterExtractor extractor) {
-    Double threshold = extractor.singleValueParameter(VALUE, Double.class);
-    String stringOperation = extractor.selectedSingleValue(OPERATION, String.class);
-
-    String operation = "GT";
-
-    if (stringOperation.equals("<=")) {
-      operation = "LE";
-    } else if (stringOperation.equals("<")) {
-      operation = "LT";
-    } else if (stringOperation.equals(">=")) {
-      operation = "GE";
-    } else if (stringOperation.equals("==")) {
-      operation = "EQ";
-    } else if (stringOperation.equals("!=")) {
-      operation = "IE";
-    }
-
-    String filterProperty = extractor.mappingPropertyValue(NUMBER_MAPPING);
-
-    ThresholdDetectionParameters staticParam = new ThresholdDetectionParameters(sepa,
-            threshold,
-            ThresholdDetectionOperator.valueOf(operation),
-            filterProperty);
-
-    return new ConfiguredEventProcessor<>(staticParam, ThresholdDetection::new);
-  }
-}
